requirements.txt
tensorflow
pandas
numpy
scikit-learn
streamlit
pillow
tqdm
selenium
beautifulsoup4
webdriver-manager
